% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Configurer.R
\name{Configurer}
\alias{Configurer}
\title{Configurer}
\description{
This class is for getting and setting configuration options. You can
set any text values you like on the fly with
setValue and retrieve them with getValue. They are not saved, but
might be useful during runtime, since the Configurer is available to
all the objects used in this package.
}
\section{Fields}{

\describe{
\item{\code{config}}{Reference to a Configurer object.}
}}

