% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Illumina450kAnnotation.R
\name{Illumina450kAnnotation}
\alias{Illumina450kAnnotation}
\title{Illumina450kAnnotation}
\description{
Provides an Illumina450kAnnotation object.
This object will hold an annotation in a memory efficient manner.
Only the data that is needed will actually be loaded.
}
\section{Fields}{

\describe{
\item{\code{configurer}}{- holds global configuration data in Configurer object}

\item{\code{dao}}{- a reference to a DAO for expression data}
}}

